+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased - 2025-XX-XX
 
+### Visual
+- Changed: Base Minimap Edits has been updated to reflect new events, boss tiles, and security rooms.
+
 ## 0.4.0 - 2025-05-15
 - Changelog Created
 - Added: 6 new Item Locations. 3 Major/Events: Boiler, Animals, Auxiliary Power. 3 Minor/Tanks: Subzero Containment, Quarantine Bay, Northeast Stabilizer